requirements.txt
git+git://github.com/learning-at-home/hivemind@calm
bitsandbytes-cuda111>=0.26.0
transformers==4.13.0
tokenizers>=0.10.2
datasets>=1.11.0
torch_optimizer>=0.1.0
wandb>=0.10.33
nltk>=3.6.2
revlib==1.3.1
sentencepiece
aiohttp
requests>=2.24.0
termcolor
